Remove -Werror, add -Wpedantic and -Wc++-compat
Newer versions of gcc/clang will certainly introduce more warnings.
By removing -Werror this helps ensure that the code will continue
to compile even as new checks are added to the compiler.

Adding -Wpedantic and other additional -W flags increases the
likelihood that a packager will become aware of a possible issue,
yet removing -Werror means the packager or developer will not be
required to resolve the possible issue immediately.